requirements.txt
git+https://github.com/Rapptz/discord-ext-menus

discord.py
xkcd-wrapper
psutil
bs4
py-cpuinfo
pyfiglet
GitPython
pygit2
pymongo
PyYAML
mechanicalsoup
tld
dnspython
pyparsing

aiosqlite
aiohttp_client_cache
aiohttp
